requirements.txt
pytest==3.2.1
numpy>=1.19.5
pandas>=1.0.0
pyroaring==0.3.3
paretoset==1.2.0
scikit-learn-extra==0.2.0
tsp-solver2==0.4.1
delayed==0.11.0b1